move bridgebutton
